+++ README.md
@@ -4,7 +4,7 @@ This program (approximately) simulates a Turing machine.
 
 It consists of several parts.
 
-You can find the webpage for creating program JSON files [at this link](https://eluni.co/TuringMachine/transition-maker.html)
+You can find the webpage for creating program JSON files [at this link](https://eluni.co/TuringMachine/web/transition-maker.html)
 
 ## importing TuringMachine
 First there is a TuringMachine class which can be imported and used in other programs.
@@ -41,7 +41,7 @@ First we will discuss the creation of the files
 #### program.json
 This file contains the states, transition rules, initial state, and initial tape index
 that your program will use. It is a JSON file with the following schema. You can use
-[this website](https://eluni.co/TuringMachine/transition-maker.html) to easily construct 
+[this website](https://eluni.co/TuringMachine/web/transition-maker.html) to easily construct 
 one of these files
 ```json
 {
